+++ packages/storybook/src/nimble/anchored-region/anchored-region.react.tsx
@@ -1,4 +1,4 @@
-import { AnchoredRegion } from '@ni/nimble-components/dist/esm/anchored-region';
+import { AnchoredRegion } from '../../../../nimble-components/src/anchored-region';
 import { wrap } from '../../utilities/react-wrapper';
 
 export const NimbleAnchoredRegion = wrap(AnchoredRegion);