+++ packages/storybook/src/nimble/mapping/text/mapping-text.react.tsx
@@ -1,4 +1,4 @@
-import { MappingText } from '@ni/nimble-components/dist/esm/mapping/text';
+import { MappingText } from '../../../../../nimble-components/src/mapping/text';
 import { wrap } from '../../../utilities/react-wrapper';
 
 export const NimbleMappingText = wrap(MappingText);