+++ packages/storybook/src/nimble/patterns/button/doc-strings.ts
@@ -1,8 +1,10 @@
-import { iconDescription as baseIconDescription } from '../../../utilities/storybook';
+import { textContentDescription as baseTextContentDescription, iconDescription as baseIconDescription } from '../../../utilities/storybook';
 
-export const appearanceDescription = 'This attribute affects the appearance of the button.';
+export const textContentDescription = (options: { componentName: string }): string => `${baseTextContentDescription(options)} This will be hidden when \`content-hidden\` is set but should always be provided; see the **Accessibility** section for more info.`;
 
-export const appearanceVariantDescription = 'This attribute has no effect on buttons with a `ghost` appearance. There is no `accent` appearance-variant for the `color` UI.';
+export const appearanceDescription = 'This attribute affects the appearance of the button. See the **Styling** section for more information.';
+
+export const appearanceVariantDescription = 'This attribute configures additional variations of the button appearance. It has no effect on buttons with a `ghost` appearance. There is no `accent` appearance-variant for the `color` UI.  See the **Styling** section for more information.';
 
 export const contentHiddenDescription = 'When set, this attribute hides the text and end icon, leaving only the start icon visible.';