+++ packages/storybook/src/nimble/radio-group/radio-group.react.tsx
@@ -1,4 +1,4 @@
-import { RadioGroup } from '@ni/nimble-components/dist/esm/radio-group';
+import { RadioGroup } from '../../../../nimble-components/src/radio-group';
 import { wrap } from '../../utilities/react-wrapper';
 
 export const NimbleRadioGroup = wrap(RadioGroup);