+++ packages/storybook/src/nimble/select/select-matrix.stories.ts
@@ -5,11 +5,11 @@ import {
     controlLabelFont,
     controlLabelFontColor,
     standardPadding
-} from '@ni/nimble-components/dist/esm/theme-provider/design-tokens';
-import { listOptionTag } from '@ni/nimble-components/dist/esm/list-option';
-import { Select, selectTag } from '@ni/nimble-components/dist/esm/select';
-import { DropdownAppearance } from '@ni/nimble-components/dist/esm/patterns/dropdown/types';
-import { waitForUpdatesAsync } from '@ni/nimble-components/dist/esm/testing/async-helpers';
+} from '../../../../nimble-components/src/theme-provider/design-tokens';
+import { listOptionTag } from '../../../../nimble-components/src/list-option';
+import { Select, selectTag } from '../../../../nimble-components/src/select';
+import { DropdownAppearance } from '../../../../nimble-components/src/patterns/dropdown/types';
+import { waitForUpdatesAsync } from '../../../../nimble-components/src/testing/async-helpers';
 import { createStory } from '../../utilities/storybook';
 import {
     createMatrixThemeStory,
@@ -39,6 +39,12 @@ const valueStates = [
 ] as const;
 type ValueState = (typeof valueStates)[number];
 
+const clearableStates = [
+    ['', false],
+    ['Clearable', true]
+] as const;
+type ClearableState = (typeof clearableStates)[number];
+
 const metadata: Meta = {
     title: 'Tests/Select',
     parameters: {
@@ -54,6 +60,7 @@ const component = (
     [appearanceName, appearance]: AppearanceState,
     [errorName, errorVisible, errorText]: ErrorState,
     [valueName, valueValue]: ValueState,
+    [clearableName, clearable]: ClearableState
 ): ViewTemplate => html`
     <div style="
         display: inline-flex;
@@ -62,11 +69,12 @@ const component = (
         font: var(${controlLabelFont.cssCustomProperty});
         color: var(${controlLabelFontColor.cssCustomProperty});"
     >
-        <label>${() => errorName} ${() => disabledName} ${() => appearanceName} ${() => valueName}</label>
+        <label>${() => errorName} ${() => disabledName} ${() => appearanceName} ${() => valueName} ${() => clearableName}</label>
         <${selectTag}
             ?error-visible="${() => errorVisible}"
             error-text="${() => errorText}"
             ?disabled="${() => disabled}"
+            ?clearable="${() => clearable}"
             appearance="${() => appearance}"
             style="width: 250px;"
         >
@@ -83,7 +91,8 @@ export const selectThemeMatrix: StoryFn = createMatrixThemeStory(
         disabledStates,
         appearanceStates,
         errorStates,
-        valueStates
+        valueStates,
+        clearableStates
     ])
 );