+++ packages/storybook/src/nimble/table-column/mapping/table-column-mapping.react.tsx
@@ -1,4 +1,4 @@
-import { TableColumnMapping } from '@ni/nimble-components/dist/esm/table-column/mapping';
+import { TableColumnMapping } from '../../../../../nimble-components/src/table-column/mapping';
 import { wrap } from '../../../utilities/react-wrapper';
 
 export const NimbleTableColumnMapping = wrap(TableColumnMapping);