+++ packages/storybook/src/nimble/toolbar/toolbar.stories.ts
@@ -1,12 +1,12 @@
 import { html } from '@microsoft/fast-element';
 import type { Meta, StoryObj } from '@storybook/html';
-import { buttonTag } from '@ni/nimble-components/dist/esm/button';
-import { iconCogTag } from '@ni/nimble-components/dist/esm/icons/cog';
-import { iconEyeTag } from '@ni/nimble-components/dist/esm/icons/eye';
-import { iconFilterTag } from '@ni/nimble-components/dist/esm/icons/filter';
-import { iconPencilTag } from '@ni/nimble-components/dist/esm/icons/pencil';
-import { iconTrashTag } from '@ni/nimble-components/dist/esm/icons/trash';
-import { toolbarTag } from '@ni/nimble-components/dist/esm/toolbar';
+import { buttonTag } from '../../../../nimble-components/src/button';
+import { iconCogTag } from '../../../../nimble-components/src/icons/cog';
+import { iconEyeTag } from '../../../../nimble-components/src/icons/eye';
+import { iconFilterTag } from '../../../../nimble-components/src/icons/filter';
+import { iconPencilTag } from '../../../../nimble-components/src/icons/pencil';
+import { iconTrashTag } from '../../../../nimble-components/src/icons/trash';
+import { toolbarTag } from '../../../../nimble-components/src/toolbar';
 import { createUserSelectedThemeStory } from '../../utilities/storybook';
 
 const metadata: Meta = {