+++ packages/storybook/src/spright/rectangle/rectangle.react.tsx
@@ -1,4 +1,4 @@
-import { Rectangle } from '@ni/spright-components/dist/esm/rectangle';
+import { Rectangle } from '../../../../spright-components/src/rectangle';
 import { wrap } from '../../utilities/react-wrapper';
 
 export const SprightRectangle = wrap(Rectangle);